run dsd integration tests on the CI
