+++ tasks/dogstatsd.py
@@ -11,7 +11,7 @@ from invoke.exceptions import Exit
 from .build_tags import get_build_tags
 from .utils import get_ldflags, bin_name, get_root
 from .utils import REPO_PATH
-
+from .go import deps
 
 # constants
 DOGSTATSD_BIN_PATH = os.path.join(".", "bin", "dogstatsd")
@@ -134,3 +134,17 @@ def omnibus_build(ctx):
             "overrides": overrides_cmd
         }
         ctx.run(cmd.format(**args))
+
+@task
+def integration_tests(ctx, install_deps=False):
+    """
+    Run integration tests for the Agent
+    """
+    if install_deps:
+        deps(ctx)
+
+    build_tags = get_build_tags()
+
+    # config_providers
+    cmd = "go test -tags '{}' {}/test/integration/dogstatsd/..."
+    ctx.run(cmd.format(" ".join(build_tags), REPO_PATH))
