+++ test/integration/dogstatsd/origin_detection_skip.go
@@ -1,0 +1,15 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2017 Datadog, Inc.
+
+// +build !linux
+
+package dogstatsd
+
+import "testing"
+
+// noop version for unsupported platforms
+func testUDSOriginDetection(t *testing.T) {
+	t.Log("Unsupported platform, skip...")
+}
